app/requirements.txt
mysql-connector-python
Flask
Flask-WTF
Flask-Login